WIP: ok endpoints

+++ src/main/java/com/pepit/model/User.java
@@ -26,7 +26,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    Long id;
+    Integer id;
 
     @Column(name = "email")
     String email;
@@ -42,7 +42,7 @@ public class User {
 
     @OneToMany
     List<Filter> filters;
-    
+
     @ManyToOne
     Company company;
